tests/submission/W3C/domhtml/HTMLImageElement09.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - IMG</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-img-element.html#dom-img-src" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<img src="/support/images/yellow-rectangle.png">
</p>



<script>
var domtest = async_test("The src attribute contains an URI designating the source of this image.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vsrc;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("img");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vsrc = testNode.src;

      assert_uri_equals(document.location.protocol,null,document.location.host,"yellow-rectangle.png",null,null,null,true,vsrc);

    domtest.done();
}));

</script>
<p>The src attribute contains an URI designating the source of this image.

    Retrieve the src attribute and examine its value.</p>
<div id="log"></div>
</body></html>
